moving of first set of internal functions helpers
